+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ## [0.1.4](https://github.com/dbalsom/martypc/releases/tag/0.1.4) (2023-07-XX)
 
+
+* Add debug_keyboard config flag - this will print keyboard event info to the console for support
 * Don't recalculate composite parameters if mode change was enable bit only
 
 ## [0.1.3](https://github.com/dbalsom/martypc/releases/tag/0.1.3) (2023-07-06)